Remove broken feature
Playback was supposed to finish and then move to the next file in the 
movieList. This didn't work because inputVideo never got changed even 
though currentVideo did. Removing this for now as movies play so long 
that it shouldn't be a problem to ssh in and start the next one once a 
year or so. Changed so it restarts the same video.